requirements.txt
joblib==1.0.1
matplotlib==3.3.4
mpose
numpy==1.19.5
optuna==2.8.0
pandas==1.1.5
PyYAML==6.0
scikit_learn==1.0.2
scipy==1.5.4
tensorflow==2.6.0
tensorflow_addons==0.14.0
tensorflow_datasets==4.4.0
